Allow apostrophes in local part of email addresses
After using this gem in production successfully we had a real user with
a valid email address that contained an apostrophe in the local part of
the email address. It was in the form
`first_name.o'last_name@domain.com` and this user was unable to be
added to the application due to the email validation. After some
investigation it appears that an apostrophe is a valid character in the
local part (part before the @ symbol).

See https://en.wikipedia.org/wiki/Email_address#Local_part for details.

There are other special characters listed here that are currently not
allowed by the validation but for now I decided just to add the
characters that we need.

I've added email addresses to the valid and invalid lists used by the
specs to ensure that apostrophes are only valid in the local part.